ytBrief_Backend/requirements.txt
flask 
flask-cors
python-dotenv
beautifulsoup4
requests
youtube-transcript-api
google-api-python-client
google-generativeai